feat: 🎸 AnimationBuilder has more type info, which improves DX
AnimationBuilder has two generics now, one of them is for AnimationType
and another one for AnimationOptions that can be used with the specified
AnimationType. It is improves a developer experience of using
AnimationBuilder a lot, since it now suggest completion for your options
e.g. and type checks other information.