feat: 🎸 generic interface AnimationObject over type and options
AnimationObject interface is now generic over type of the animation and
the options the animation of the type can accept. It is simplifies
implementing new animations where you can specify your type of
AnimationObject just using generic one. Also, it adds a possibility for
TypeScript to infer some of the types when using these interfaces.