feat: 🎸 SlideBuilder can be started with predefined shapes\anim
SlideBuilder.start() method accepts now two arguments: a Record with
shape names as keys and ShapeRenderable as values and a Record with
animation names as keys and Animationable as values. These are mixing
into slide instance when instantiating a builder instance. Also, it
improves type information and allows to make a predefined set of
shapes\animations that can be shared amongst different SlideBuilders.