fix: 🐛 issue with "Untitled Slide" already exists in builder
When adding new slides via withSlide() in DeckBuilder and do not provide
names for these slides, it uses "Untitled Slide" by default. The problem
is that when you adding more than one such slide, it throws an error
that slide already exists. This commit fixes it by introducing a random
suffix to "Untitled Slide" name.